Add `page_data` helper to standardise page titles, headings and errors
This will help ensure consistency across views.

`page_title`  follows the regular format of the title of the current page followed by the service name.

When `errors: true` the title will be prefixed with `Error: `.  It is expected that this is usually set to `errors: @the_form_object.errors.any?`

`page_header` defaults to the title text but can be overridden if the two differ. The size is controllable using `header_size`.